backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyJWT==2.8.0
Werkzeug==3.0.3
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
pymongo==4.6.1
